+++ telemetry/testing.go
@@ -1,16 +1,10 @@
 package telemetry
 
 import (
-	"context"
-	"log/slog"
-	"strconv"
-
 	noopmetric "go.opentelemetry.io/otel/metric/noop"
 	nooptrace "go.opentelemetry.io/otel/trace/noop"
 
 	"github.com/dogmatiq/spruce"
-	"go.opentelemetry.io/otel/log"
-	"go.opentelemetry.io/otel/log/embedded"
 )
 
 // TestingT is the subset of [testing.TB] that is used by the test provider.
@@ -26,141 +20,8 @@ func NewTestProvider(t TestingT) *Provider {
 	return &Provider{
 		TracerProvider: nooptrace.NewTracerProvider(),
 		MeterProvider:  noopmetric.NewMeterProvider(),
-		LoggerProvider: &testLoggerProvider{
-			logger: spruce.NewTestLogger(t),
+		LoggerProvider: &slogProvider{
+			Target: spruce.NewTestLogger(t),
 		},
-	}
-}
-
-type testLoggerProvider struct {
-	embedded.LoggerProvider
-	logger *slog.Logger
-}
-
-func (p *testLoggerProvider) Logger(name string, _ ...log.LoggerOption) log.Logger {
-	return &testLogger{logger: p.logger.WithGroup(name)}
-}
-
-type testLogger struct {
-	embedded.Logger
-	logger *slog.Logger
-}
-
-func (l *testLogger) Emit(ctx context.Context, rec log.Record) {
-	level := slog.LevelDebug
-
-	switch {
-	case rec.Severity() >= log.SeverityError:
-		level = slog.LevelError
-	case rec.Severity() >= log.SeverityWarn:
-		level = slog.LevelWarn
-	case rec.Severity() >= log.SeverityInfo:
-		level = slog.LevelInfo
-	}
-
-	var attrs []slog.Attr
-
-	message := "?"
-
-	if rec.Body().Kind() == log.KindString {
-		message = rec.Body().AsString()
-	} else if !rec.Body().Empty() {
-		attrs = append(
-			attrs,
-			convertValue("body", rec.Body()),
-		)
-	}
-
-	if ev := rec.EventName(); ev != "" {
-		attrs = append(attrs, slog.String("event", ev))
-	}
-
-	rec.WalkAttributes(
-		func(kv log.KeyValue) bool {
-			attrs = append(
-				attrs,
-				convertValue(kv.Key, kv.Value),
-			)
-			return true
-		},
-	)
-
-	if !rec.Timestamp().IsZero() {
-		attrs = append(attrs, slog.Time("timestamp", rec.Timestamp()))
-	}
-
-	if !rec.ObservedTimestamp().IsZero() {
-		attrs = append(attrs, slog.Time("observed_timestamp", rec.ObservedTimestamp()))
-	}
-
-	l.logger.LogAttrs(
-		ctx,
-		level,
-		message,
-		attrs...,
-	)
-}
-
-func (l *testLogger) Enabled(context.Context, log.EnabledParameters) bool {
-	return true
-}
-
-func convertValue(name string, v log.Value) slog.Attr {
-	switch v.Kind() {
-	case log.KindEmpty:
-		return slog.Any(name, nil)
-
-	case log.KindBool:
-		return slog.Bool(name, v.AsBool())
-
-	case log.KindFloat64:
-		return slog.Float64(name, v.AsFloat64())
-
-	case log.KindInt64:
-		return slog.Int64(name, v.AsInt64())
-
-	case log.KindString:
-		return slog.String(name, v.AsString())
-
-	case log.KindBytes:
-		return slog.Any(name, v.AsBytes())
-
-	case log.KindSlice:
-		var attrs []slog.Attr
-		for i, elem := range v.AsSlice() {
-			attrs = append(
-				attrs,
-				convertValue(
-					strconv.Itoa(i),
-					elem,
-				),
-			)
-		}
-		return groupAttrs(name, attrs)
-
-	case log.KindMap:
-		var attrs []slog.Attr
-		for _, pair := range v.AsMap() {
-			attrs = append(
-				attrs,
-				convertValue(
-					pair.Key,
-					pair.Value,
-				),
-			)
-		}
-		return groupAttrs(name, attrs)
-
-	default:
-		return slog.String(name, v.String())
-	}
-}
-
-// TODO: remove this when go.mod is updated to Go 1.25 or later, then use
-// [slog.GroupAttrs] instead.
-func groupAttrs(name string, attrs []slog.Attr) slog.Attr {
-	return slog.Attr{
-		Key:   name,
-		Value: slog.GroupValue(attrs...),
 	}
 }